+++ lazyllm/tools/rag/readers/ocrReader/ocr_reader_base.py
@@ -4,8 +4,8 @@ from enum import Enum
 
 from ..readerBase import _RichReader
 from ...doc_node import DocNode
-from .ocr_ir import Block
-
+from .ocr_ir import Block, Cell
+from lazyllm.thirdparty import bs4
 
 class ServiceVariant(str, Enum):
     ONLINE = 'online'
@@ -47,6 +47,32 @@ class _OcrReaderBase(_RichReader):
         response_raw_text = self._fetch_response(file, use_cache=use_cache)
         return self._build_nodes_from_response(response_raw_text, file, extra_info)
 
+    @staticmethod
+    def _parse_table_html( html_text: str) -> List[Cell]:
+        soup = bs4.BeautifulSoup(html_text, 'html.parser')
+        cells: List[Cell] = []
+        for row_idx, tr in enumerate(soup.find('table').find_all('tr')):
+            for col_idx, td in enumerate(tr.find_all(['td', 'th'])):
+                cells.append(Cell(
+                    row=row_idx,
+                    col=col_idx,
+                    rowspan=int(td.get('rowspan', 1)),
+                    colspan=int(td.get('colspan', 1)),
+                    text=td.get_text(strip=True),
+                ))
+        return cells
+
+    @staticmethod
+    def _first(val):
+        if isinstance(val, list) and val:
+            return val[0]
+        if isinstance(val, str):
+            return val
+        return None
+
+    @staticmethod
+    def _make_anchor(text: str) -> str:
+        return text.strip().replace(' ', '-').replace('\n', '-')[:64]
 
 class _Adapter:
     def _adapt_json_to_IR(self, raw: dict) -> List[Block]:
